Actually fixed the trailing comment bug
The way I implemented adding the EOF Token to the end of the scanner
was prety hacky, and it was actualy the cause of escaping the loop in
parse(). I reworked that part of the scanner, and it also fixed the
issue where the scanner would keep functioning after an error was
emitted. Now I don't need to check for an empty scanner, and trailing
comments don't eat the EOF.